% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimand.R
\docType{data}
\name{ESTIMAND_STRATEGIES}
\alias{ESTIMAND_STRATEGIES}
\title{Valid ICH E9(R1) intercurrent-event strategies}
\format{
An object of class \code{character} of length 5.
}
\usage{
ESTIMAND_STRATEGIES
}
\description{
Valid ICH E9(R1) intercurrent-event strategies
}
\keyword{datasets}
