% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioClinStats-package.R
\docType{package}
\name{PhysioClinStats-package}
\alias{PhysioClinStats}
\alias{PhysioClinStats-package}
\title{PhysioClinStats: clinical inference engine}
\description{
Mixed-effects / MMRM longitudinal models, single-case (N-of-1) designs,
estimands with multiple imputation, causal mediation, declared target-trial
emulation, and per-estimate uncertainty. Heavy modelling backends are
optional, guarded Suggests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioClinStats}
  \item \url{https://x-biosignal.github.io/PhysioClinStats/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioExperiment/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
